separation of responsibilities for math operations
